requirements.txt
# Telegram-бот
aiogram==3.13.0

# HTTP-клиент (асинхронный)
httpx==0.27.2

# OpenAI GPT
openai==1.51.2

# Планировщик задач
APScheduler==3.10.4

# Переменные окружения и конфигурация
python-dotenv==1.0.1
pydantic-settings==2.5.2
pydantic>=2.4.1,<2.9

# Работа с файлами async
aiofiles==23.2.1

# Часовые пояса
pytz==2024.2
